+++ src/store/slices/trackedTransactions/trackedTransactionsSlice.ts
@@ -2,7 +2,7 @@ import { StateCreator } from 'zustand/vanilla';
 import { StoreType, MutatorsIn } from 'store/store.types';
 import { TrackedTransactionsSliceType } from './trackedTransactionsSlice.types';
 
-const initialState: TrackedTransactionsSliceType = {};
+export const initialState: TrackedTransactionsSliceType = {};
 
 function getTrackedTransactionsSlice(): StateCreator<
   StoreType,