+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added format amount controller](https://github.com/multiversx/mx-sdk-dapp-core/pull/65)
 - [Added logout check on forced address change](https://github.com/multiversx/mx-sdk-dapp-core/pull/64)
 - [Refactor LedgerProviderStrategy](https://github.com/multiversx/mx-sdk-dapp-core/pull/63)
 - [Added react hooks](https://github.com/multiversx/mx-sdk-dapp-core/pull/62)